Conflicts:
	scripts/commands/send.lua
	scripts/globals/abilities/meditate.lua
	scripts/globals/abilities/sentinel.lua
	sql/zone_settings.sql
	src/map/lua/lua_baseentity.cpp

+++ src/map/map.cpp
@@ -1080,7 +1080,7 @@ int32 map_config_read(const int8* cfgName)
         }
         else if (strcmp(w1,"nm_hp_multiplier") == 0)
         {
-            map_config.nm_hp_multiplier = atoi(w2);
+            map_config.nm_hp_multiplier = atof(w2);
         }
         else if (strcmp(w1,"mob_hp_multiplier") == 0)
         {
@@ -1092,7 +1092,7 @@ int32 map_config_read(const int8* cfgName)
         }
         else if (strcmp(w1,"nm_mp_multiplier") == 0)
         {
-            map_config.nm_mp_multiplier = atoi(w2);
+            map_config.nm_mp_multiplier = atof(w2);
         }
         else if (strcmp(w1,"mob_mp_multiplier") == 0)
         {
